[Comm] Extract the CuTe DSL AR fusion core out of the Qwen module
The fusion mechanism that landed with the Qwen3.5 stack is almost entirely
architecture-agnostic: the deferred-MoE handoff record, the process-local
workspace handle, and the LayerCommunicator hooks that route an eligible
layer into the fused collective. Exactly one thing in it is Qwen-specific --
`gemma_weight`, because Qwen3.5 normalizes with GemmaRMSNorm while the
DeepSeek-V3 family uses a plain RMSNorm.

Move the generic parts to layers/moe/cutedsl_ar_fusion.py and reduce the
per-architecture surface to one class attribute, NORM_WEIGHT_ATTR, read
through a _norm_gamma() hook at the four sites that touched gemma_weight
directly. qwen35_flashinfer_fusion.py stays as the Qwen flavour plus
back-compat aliases, so qwen2_moe, qwen3_5 and the registered Qwen test
import exactly what they did before.

Two things that looked Qwen-specific but are not: weight_bias=0.0 is right
for both flavours (the kernel computes x * (gamma + weight_bias), and
neither wants a bias), and consuming the final layer's handoff with the
model's own final norm is generic behaviour.
